requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
Pillow>=11.0
django-cors-headers>=4.3

